requirements.txt
Django==4.2.19
djangorestframework==3.15.2
dj-database-url==2.3.0
django-environ==0.8.1
django-crispy-forms==2.0
crispy-bootstrap5==0.6


psycopg2-binary==2.9.10


asgiref==3.8.1
gunicorn==23.0.0
sqlparse==0.5.3
typing-extensions==4.12.2
packaging==24.2
autopep8==1.6.0 
certifi==2024.8.30
filelock==3.16.0 
platformdirs==4.3.2
distlib==0.3.8

django-login-required-middleware==0.9.0
django-tracing==1.0.6
django-fsm==2.8.1

django-admin-soft-dashboard==1.0.11
opencv-python-headless

